refactor(queue): collapse retain queue path to queue store + delivery
Removes the lifecycle/retain-queue pass-through wrapper and folds
retainDurably/buildDurableRetainJob into lifecycle/retain.ts. Queue
config-level helpers (enqueueRetain, flushRetain, readQueuedRetains,
removeQueuedRetains, summarizeRetain) now live next to the store in
queue/queue.ts, so lifecycle policy talks to exactly two queue modules:
the store and the delivery engine. Durability semantics are unchanged.

Refs #412
