+++ extensions/lifecycle/retain.ts
@@ -2,13 +2,15 @@ import type { AgentEndEvent } from "@earendil-works/pi-coding-agent";
 import type {
   HindsightCapabilities,
   HindsightLikeClient,
+  HindsightObservationScopes,
   ResolvedConfig,
   RetainJob,
+  UpdateMode,
 } from "../types.js";
 import { baseTags } from "../banks/banking.js";
 import { projectMessages } from "../utils/messages.js";
 import { contextLabel, liveDocumentId, stableSessionId } from "../utils/session.js";
-import { enqueueRetain, flushRetain } from "./retain-queue.js";
+import { enqueueRetain, flushRetain, summarizeRetain } from "../queue/queue.js";
 import { createMemoryIdentity } from "../operations/memory-identity.js";
 import { expandObservationScopes } from "./observation-scopes.js";
 import { buildRetainJob as buildRetainJobCore } from "./retain-job-builder.js";
@@ -87,4 +89,72 @@ export async function enqueueRetainFromAgentEnd(args: {
     }
   }
   return { queued: true, sent: result.sent, remaining: result.remaining };
+}
+
+export type DurableRetainSource = "auto" | "tool" | "command" | "import";
+
+export interface RetainDurablyArgs {
+  cwd: string;
+  config: ResolvedConfig;
+  client: HindsightLikeClient;
+  bankId: string;
+  content: string;
+  context: string;
+  tags: string[];
+  documentId: string;
+  updateMode: UpdateMode;
+  metadata?: Record<string, string>;
+  source: DurableRetainSource;
+  timestamp?: string;
+  observationScopes?: HindsightObservationScopes;
+  documentTags?: string[];
+  entities?: RetainJob["item"]["entities"];
+  async?: boolean;
+  capabilities?: HindsightCapabilities;
+}
+
+export interface RetainDurablyResult {
+  enqueued: boolean;
+  sent: number;
+  remaining: number;
+  deadLettered: number;
+  bankId: string;
+  documentId: string;
+  queueJobId: string;
+  updateMode: UpdateMode;
+  operationIds?: string[];
+}
+
+export function buildDurableRetainJob(args: Omit<RetainDurablyArgs, "client">): RetainJob {
+  return buildRetainJobCore({
+    ...args,
+    metadata: {
+      ...args.metadata,
+      source: "pi-hindsight",
+      retainSource: args.source,
+    },
+  });
+}
+
+export async function retainDurably(args: RetainDurablyArgs): Promise<RetainDurablyResult> {
+  const job = buildDurableRetainJob(args);
+  const receipt = {
+    bankId: job.bankId,
+    documentId: job.documentId,
+    queueJobId: job.id,
+    updateMode: job.updateMode,
+  };
+  const enqueueResult = await enqueueRetain(args.cwd, args.config, job);
+  if (enqueueResult.previousLength > 0) {
+    const summary = await summarizeRetain(args.cwd, args.config);
+    return {
+      ...receipt,
+      enqueued: true,
+      sent: 0,
+      remaining: summary.active.valid + summary.active.malformed,
+      deadLettered: 0,
+    };
+  }
+  const result = await flushRetain(args.cwd, args.config, args.client, { maxJobs: 1 });
+  return { ...receipt, enqueued: true, ...result };
 }
